mesh: Normalize OOB private key to match spec
The Mesh Profile specification indicates that Public ECC keys are sent
Over-The-Air in Big Endian format, and in the sample data both the
Private and resulting Public keys are also listed in this format.

Our internal calculations are done in Little Endian, so a translation
must be made. For simplicity, we need to make make sure that our DBus
apis, and our debugging output match what the specification uses, so
these changes Normalize the Public and Private keys to Little Endian
when calculating shared secrets, and convert back to Big Endian when
shared over debugging and API interfaces.